+++ apps/explorer/config/runtime/test.exs
@@ -1,5 +1,7 @@
 import Config
 
+alias EthereumJSONRPC.Variant
+
 config :explorer, Explorer.ExchangeRates, enabled: false, store: :ets, fetch_btc_value: true
 
 config :explorer, Explorer.Chain.Cache.BlockNumber, enabled: false
@@ -26,15 +28,7 @@ config :explorer, Explorer.Staking.ContractState, enabled: false
 config :explorer,
   realtime_events_sender: Explorer.Chain.Events.SimpleSender
 
-variant =
-  if is_nil(System.get_env("ETHEREUM_JSONRPC_VARIANT")) do
-    "parity"
-  else
-    System.get_env("ETHEREUM_JSONRPC_VARIANT")
-    |> String.split(".")
-    |> List.last()
-    |> String.downcase()
-  end
+variant = Variant.get()
 
 Code.require_file("#{variant}.exs", "#{__DIR__}/../../../explorer/config/test")
 Code.require_file("#{variant}.exs", "#{__DIR__}/../../../indexer/config/test")