+++ apps/explorer/config/test/nethermind.exs
@@ -5,10 +5,10 @@ config :explorer,
   json_rpc_named_arguments: [
     transport: EthereumJSONRPC.Mox,
     transport_options: [],
-    variant: EthereumJSONRPC.Parity
+    variant: EthereumJSONRPC.Nethermind
   ],
   subscribe_named_arguments: [
     transport: EthereumJSONRPC.Mox,
     transport_options: [],
-    variant: EthereumJSONRPC.Parity
+    variant: EthereumJSONRPC.Nethermind
   ]