+++ apps/indexer/lib/indexer/block/fetcher.ex
@@ -29,7 +29,7 @@ defmodule Indexer.Block.Fetcher do
     UncleBlock
   }
 
-  alias Indexer.Tracer
+  alias Indexer.{Prometheus, Tracer}
 
   alias Indexer.Transform.{
     AddressCoinBalances,
@@ -121,6 +121,9 @@ defmodule Indexer.Block.Fetcher do
         _.._ = range
       )
       when callback_module != nil do
+    {fetch_time, fetched_blocks} =
+      :timer.tc(fn -> EthereumJSONRPC.fetch_blocks_by_range(range, json_rpc_named_arguments) end)
+
     with {:blocks,
           {:ok,
            %Blocks{
@@ -128,15 +131,18 @@ defmodule Indexer.Block.Fetcher do
              transactions_params: transactions_params_without_receipts,
              block_second_degree_relations_params: block_second_degree_relations_params,
              errors: blocks_errors
-           }}} <- {:blocks, EthereumJSONRPC.fetch_blocks_by_range(range, json_rpc_named_arguments)},
+           }}} <- {:blocks, fetched_blocks},
          blocks = TransformBlocks.transform_blocks(blocks_params),
          {:receipts, {:ok, receipt_params}} <- {:receipts, Receipts.fetch(state, transactions_params_without_receipts)},
          %{logs: logs, receipts: receipts} = receipt_params,
          transactions_with_receipts = Receipts.put(transactions_params_without_receipts, receipts),
          %{token_transfers: token_transfers, tokens: tokens} = TokenTransfers.parse(logs),
          %{mint_transfers: mint_transfers} = MintTransfers.parse(logs),
+         %FetchedBeneficiaries{params_set: beneficiary_params_set, errors: beneficiaries_errors} =
+           fetch_beneficiaries(blocks, transactions_with_receipts, json_rpc_named_arguments),
          addresses =
            Addresses.extract_addresses(%{
+             block_reward_contract_beneficiaries: MapSet.to_list(beneficiary_params_set),
              blocks: blocks,
              logs: logs,
              mint_transfers: mint_transfers,
@@ -145,6 +151,7 @@ defmodule Indexer.Block.Fetcher do
            }),
          coin_balances_params_set =
            %{
+             beneficiary_params: MapSet.to_list(beneficiary_params_set),
              blocks_params: blocks,
              logs_params: logs,
              transactions_params: transactions_with_receipts
@@ -156,6 +163,8 @@ defmodule Indexer.Block.Fetcher do
              blocks: blocks
            }
            |> AddressCoinBalancesDaily.params_set(),
+         beneficiaries_with_gas_payment =
+           beneficiaries_with_gas_payment(blocks, beneficiary_params_set, transactions_with_receipts),
          address_token_balances = AddressTokenBalances.params_set(%{token_transfers_params: token_transfers}),
          {:ok, inserted} <-
            __MODULE__.import(
@@ -167,64 +176,14 @@ defmodule Indexer.Block.Fetcher do
                address_token_balances: %{params: address_token_balances},
                blocks: %{params: blocks},
                block_second_degree_relations: %{params: block_second_degree_relations_params},
-               block_rewards: %{errors: [], params: []},
+               block_rewards: %{errors: beneficiaries_errors, params: beneficiaries_with_gas_payment},
                logs: %{params: logs},
                token_transfers: %{params: token_transfers},
                tokens: %{on_conflict: :nothing, params: tokens},
                transactions: %{params: transactions_with_receipts}
              }
            ) do
-      Task.async(fn ->
-        %FetchedBeneficiaries{params_set: beneficiary_params_set, errors: beneficiaries_errors} =
-          fetch_beneficiaries(blocks, transactions_with_receipts, json_rpc_named_arguments)
-
-        addresses_from_block_rewards =
-          Addresses.extract_addresses(%{
-            block_reward_contract_beneficiaries: MapSet.to_list(beneficiary_params_set)
-          })
-
-        coin_balances_params_set_from_block_rewards =
-          %{
-            beneficiary_params: MapSet.to_list(beneficiary_params_set)
-          }
-          |> AddressCoinBalances.params_set()
-
-        coin_balances_params_daily_set_from_block_rewards =
-          %{
-            coin_balances_params: coin_balances_params_set_from_block_rewards,
-            blocks: blocks
-          }
-          |> AddressCoinBalancesDaily.params_set()
-
-        beneficiaries_with_gas_payment =
-          beneficiaries_with_gas_payment(blocks, beneficiary_params_set, transactions_with_receipts)
-
-        insert_params = %{
-          addresses: %{params: addresses_from_block_rewards},
-          address_coin_balances: %{params: coin_balances_params_set_from_block_rewards},
-          blocks: %{params: []},
-          block_rewards: %{errors: beneficiaries_errors, params: beneficiaries_with_gas_payment}
-        }
-
-        %MapSet{map: map} = coin_balances_params_daily_set_from_block_rewards
-
-        insert_params =
-          if map_size(map) == 0 do
-            insert_params
-          else
-            insert_params
-            |> Map.put(:address_coin_balances_daily, %{params: coin_balances_params_daily_set_from_block_rewards})
-          end
-
-        {:ok, inserted_from_rewards} =
-          __MODULE__.import(
-            state,
-            insert_params
-          )
-
-        update_addresses_cache(inserted_from_rewards[:addresses])
-      end)
-
+      Prometheus.Instrumenter.block_batch_fetch(fetch_time, callback_module)
       result = {:ok, %{inserted: inserted, errors: blocks_errors}}
       update_block_cache(inserted[:blocks])
       update_transactions_cache(inserted[:transactions])
@@ -276,7 +235,11 @@ defmodule Indexer.Block.Fetcher do
         }
       )
 
-    callback_module.import(state, options_with_broadcast)
+    {import_time, result} = :timer.tc(fn -> callback_module.import(state, options_with_broadcast) end)
+
+    no_blocks_to_import = length(options_with_broadcast.blocks.params)
+    Prometheus.Instrumenter.block_import(import_time / no_blocks_to_import, callback_module)
+    result
   end
 
   def async_import_token_instances(%{token_transfers: token_transfers}) do