+++ config/runtime/dev.exs
@@ -1,5 +1,8 @@
 import Config
 
+alias EthereumJSONRPC.Variant
+alias Explorer.Repo.ConfigHelper
+
 ######################
 ### BlockScout Web ###
 ######################
@@ -42,15 +45,10 @@ config :block_scout_web, BlockScoutWeb.Endpoint,
 database = if System.get_env("DATABASE_URL"), do: nil, else: "explorer_dev"
 hostname = if System.get_env("DATABASE_URL"), do: nil, else: "localhost"
 
-database_api_url =
-  if System.get_env("DATABASE_READ_ONLY_API_URL"),
-    do: System.get_env("DATABASE_READ_ONLY_API_URL"),
-    else: System.get_env("DATABASE_URL")
-
 pool_size =
   if System.get_env("DATABASE_READ_ONLY_API_URL"),
-    do: String.to_integer(System.get_env("POOL_SIZE", "30")),
-    else: String.to_integer(System.get_env("POOL_SIZE", "40"))
+    do: ConfigHelper.get_db_pool_size("30"),
+    else: ConfigHelper.get_db_pool_size("40")
 
 # Configure your database
 config :explorer, Explorer.Repo,
@@ -62,18 +60,12 @@ config :explorer, Explorer.Repo,
 database_api = if System.get_env("DATABASE_READ_ONLY_API_URL"), do: nil, else: database
 hostname_api = if System.get_env("DATABASE_READ_ONLY_API_URL"), do: nil, else: hostname
 
-pool_size_api = String.to_integer(System.get_env("POOL_SIZE_API", "10"))
-
 # Configure API database
 config :explorer, Explorer.Repo.Replica1,
   database: database_api,
   hostname: hostname_api,
-  url: database_api_url,
-  pool_size: pool_size_api
-
-database_account_url = System.get_env("ACCOUNT_DATABASE_URL") || System.get_env("DATABASE_URL")
-
-pool_size_account = String.to_integer(System.get_env("ACCOUNT_POOL_SIZE", "10"))
+  url: ConfigHelper.get_api_db_url(),
+  pool_size: ConfigHelper.get_api_db_pool_size("10")
 
 database_account = if System.get_env("ACCOUNT_DATABASE_URL"), do: nil, else: database
 hostname_account = if System.get_env("ACCOUNT_DATABASE_URL"), do: nil, else: hostname
@@ -82,18 +74,10 @@ hostname_account = if System.get_env("ACCOUNT_DATABASE_URL"), do: nil, else: hos
 config :explorer, Explorer.Repo.Account,
   database: database_account,
   hostname: hostname_account,
-  url: database_account_url,
-  pool_size: pool_size_account
+  url: ConfigHelper.get_account_db_url(),
+  pool_size: ConfigHelper.get_account_db_pool_size("10")
 
-variant =
-  if is_nil(System.get_env("ETHEREUM_JSONRPC_VARIANT")) do
-    "ganache"
-  else
-    System.get_env("ETHEREUM_JSONRPC_VARIANT")
-    |> String.split(".")
-    |> List.last()
-    |> String.downcase()
-  end
+variant = Variant.get()
 
 Code.require_file("#{variant}.exs", "apps/explorer/config/dev")